Solution to Check MSB of a Number

+++ Check MSB of a Number@@67882a731b4392d8fc1577e9/code.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 
 int main() {
-    int a;
+    long int a;
     scanf("%d", &a);
     a=a>>1;
     if (a&1==1) {
